k256: use `U256` as the internal representation of `Scalar` (#408)
Converts `k256::Scalar` to use `crypto_bigint::U256` as its internal
representation. It's the `k256` counterpart to the similar `p256` change
in #408.

There are a number of places where previous array-of-`u32`/`u64`
arithmetic has been retained where `U256` could potentially replace it.
However, this change is intended to just change the internal
representation while being careful not to break the existing arithmetic
implementation.
