+++ k256/Cargo.toml
@@ -49,8 +49,8 @@ keccak256 = ["digest", "sha3"]
 pem = ["elliptic-curve/pem", "pkcs8"]
 pkcs8 = ["elliptic-curve/pkcs8"]
 sha256 = ["digest", "sha2"]
-test-vectors = ["hex-literal"]
 std = ["ecdsa-core/std", "elliptic-curve/std"] # TODO: use weak activation for `ecdsa-core/std` when available
+test-vectors = ["hex-literal"]
 
 [package.metadata.docs.rs]
 features = ["ecdh", "ecdsa", "sha256", "keccak256"]
